backend-v2/requirements.txt
# ROMS V2 Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0

# Data Validation & Parsing
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.1

# Scheduling & Background Tasks
apscheduler==3.10.4
celery==5.3.4  # Optional: For heavy background processing
redis==5.0.1   # Optional: Task queue

# Email Scraping
imapclient==3.0.1
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
email-validator==2.1.0

# Web Scraping
beautifulsoup4==4.12.2
lxml==4.9.3
playwright==1.40.0
selenium==4.15.2

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytz==2023.3

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

